requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
pandas==2.0.3
PyMySQL==1.1.0
cryptography==41.0.7
